fix: DefaultRealmModule conflict
* Fix - DefaultRealmModule conflict

This commit fixes build issues when DGC-SDK is integrated into an app project, that uses Realm with default settings for other data management.

-----------------

Details :
Realms default behavior is indeed to automatically create a RealmModule called DefaultRealmModule which contains all classes extending RealmObject in a project. This module is automatically known by Realm.

Such default behavior is problematic when combining a library project & an app project that both use Realm with default settings. This is because the DefaultRealmModule will be created for both the library project and the app project, which will cause the project to fail because of duplicate class definition errors.

-----------------

Fix description :
It avoids the DefaultRealmModule conflict by setting an explicit library module for DGC-SDK (VerificaC19sdkLibraryModule). This disables the generation of the DefaultRealmModule for the library project and allows the library to be included in the app project that also uses Realm.

Issue Reference :
https://github.com/ministero-salute/it-dgc-verificac19-sdk-android/issues/90

* VerificaC19Sdk RealmModule rearrangement

- Renamed class VerificaC19sdkLibraryModule -> VerificaC19sdkRealmModule

- Moved out the RealmModule declaration outside the RevokePass class

- Reverted data/local import wildcard in VerificationViweModel & VerifierRepositoryImpl